Controller API Haciendo la Api de Controladores de documentos solicitar y de Clientes

+++ app/Http/Controllers/ClienteController.php
@@ -12,6 +12,7 @@ use Api\Reserva;
 use Session;
 use Redirect;
 use Carbon\Carbon;
+use Log;
 class ClienteController extends Controller
 {
     /**
@@ -21,21 +22,12 @@ class ClienteController extends Controller
      */
     public function index()
     {
-        $cliente = Cliente::all();       
-        
+        $cliente = Cliente::all();               
         return $cliente;
+
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
+   
     /**
      * Store a newly created resource in storage.
      *
@@ -45,61 +37,46 @@ class ClienteController extends Controller
     public function store(Request $request)
     {
         //
-        $direccion = new Direccion;
-        $direccion->pais = $request->pais;
-        $direccion->estado = $request->estado;
-        $direccion->municipio = $request->municipio;
-        $direccion->ciudad = $request->ciudad;
-        $direccion->colonia  = $request->colonia;
-        $direccion->calle = $request->calle;
-        $direccion->numeroEx = $request->exterior;
-        $direccion->numeroInt = $request->interior;
-        $direccion->codigoPostal = $request->codigo;
-        $direccion->save(); 
+        try{
+              $direccion = Direccion::create($request->all());           
+              $cliente = Cliente::create($request->all()); 
 
-        $cliente = new Cliente;
-        $cliente->nombre = $request->nombre;
-        $cliente->apellido1 = $request->apellido1;
-        $cliente->apellido2 = $request->apellido2;
-        $cliente->genero = $request->genero;
-        $cliente->email = $request->email;      
-        
-        $cliente->ultimaConexion = Carbon::now();       
-        $cliente->ip = $_SERVER['REMOTE_ADDR']; 
+              $cliente->direccion()->associate($direccion); 
+              $cliente->save();  
 
-        $cliente->direccion()->associate($direccion);               
-        $cliente->save();
-        
-        
-        for ($i=1; $i <3 ; $i++) 
-        {       
+            for ($i=1; $i <3 ; $i++) 
+            {       
             
             
-            if($request->Input('telefono_tipo'.$i)!= null)
-            {           
-                $telefono= new Telefono;
-                $telefono->tipo = $request->Input('telefono_tipo'.$i);      
-                $telefono->pais = $request->Input('telefono_pais'.$i);          
-                $telefono->area = $request->Input('telefono_area'.$i);          
-                $telefono->numero = $request->Input('telefono_numero'.$i);
+                if($request->Input('telefono_tipo'.$i)!= null)
+                {           
+                    $telefono= new Telefono([
+                        'tipo' => $request->Input('telefono_tipo'.$i),      
+                        'pais' => $request->Input('telefono_pais'.$i),          
+                        'area' => $request->Input('telefono_area'.$i),          
+                        'numero' => $request->Input('telefono_numero'.$i),
+                        ]);
 
-                $telefono->cliente()->associate($cliente);
-                $telefono->save();
-            }           
+                        $telefono->cliente()->associate($cliente);
+                        $telefono->save();
+                    
+                }           
             
+            } 
+
+
+            //$usuario = Usuario::find(Auth::user()->id);
+            //$usuario->cliente()->attach($cliente->id);
+            
+               
+            return response()->json(['status'=>true, 'message'=>'Muchas Gracias'], 200);
         }
-        
-        
+        catch(\Exception $e)
+        {
+            Log::critical("No se puede agregar el Cliente:{$e->getCode()}, {$e->getLine()}, {$e->getMessage()} ");
+            return response("Alguna cosa esta mal", 500);
+        }
 
-
-        $usuario = Usuario::find(Auth::user()->id);
-        $usuario->cliente()->attach($cliente->id);
-        
-        
-        
-
-        Session::flash('messageok','El cliente ha sido creado correctamente');
-        return Redirect::to('clientes');
 
     }
 
@@ -111,87 +88,19 @@ class ClienteController extends Controller
      */
     public function show($id)
     {
-        $cliente = Cliente::find($id);
-
-        $paquetes = $cliente->paquete;
-        
-
-
-        $info = array();
-        
-        $reservat = array();
-
-        foreach ($paquetes as $paq) 
-        {
-            $reserva = array();
-
-            $person = $paq->maximoAdulto + $paq->maximoNino;
-            $personreserva = $paq->pivot->cantAdulto + $paq->pivot->cantidadMenores;
-
-            $monto = $paq->precio;
-            
-            if($person < $personreserva && $paq->costosPersonaAdicional > 0)
-            {
-                $monto += ($personreserva-$person)*($paq->costosPersonaAdicional);
-            }
-            
-            $cllegada = new Carbon($paq->pivot->fechaLlegada);
-            $csalida = new Carbon($paq->pivot->fechaSalida);
-            $dias = $csalida->diffInDays($cllegada);
-
-            if($dias > $paq->cantidadDias && $paq->costoAdicional > 0)
-            {
-                $monto += ($dias - $paq->cantidadDias) * $paq->costoAdicional;
-            }
-            
-            $cllegada = Carbon::createFromFormat('Y-m-d H:i:s', $paq->pivot->fechaLlegada)->format('d-m-Y');
-            $ccreate = Carbon::createFromFormat('Y-m-d H:i:s', $paq->pivot->created_at)->format('d-m-Y');
-            $salida = Carbon::createFromFormat('Y-m-d H:i:s', $paq->pivot->fechaSalida)->format('d-m-Y');
-                        
-            array_push($reserva, $paq->pivot->id);
-            array_push($reserva, $ccreate); 
-            array_push($reserva, $cllegada);
-            array_push($reserva, $paq->hoteles[0]->direccion->ciudad);
-            array_push($reserva, $paq->hoteles[0]->nombre);
-            array_push($reserva, $paq->nombre);
-            array_push($reserva, $monto);
-            array_push($reserva, $dias);
-            array_push($reserva, $salida);
-            array_push($reserva, $paq->pivot->cantAdulto);
-            array_push($reserva, $paq->pivot->cantidadMenores);
-            array_push($reserva, $paq->id);     
-
-
-            array_push($reservat, $reserva);
-
+        try{
+            $cliente = Cliente::find($id);
+            if (!$cliente) {
+                return response("No existe el Cliente", 404);
+            }            
+            return response()->json($cliente, 200);
         }
+        catch(\Exception $e)
+        {
+            Log::critical("No se puede mostrar el Cliente:{$e->getCode()}, {$e->getLine()}, {$e->getMessage()} ");
+            return response("Alguna cosa esta mal", 500);
+        }        
 
-        
-        
-        
-        array_push($info,$cliente);
-
-        array_push($info, $reservat);
-
-
-
-        return response()->json($info,200);
-
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        $cliente = Cliente::find($id);              
-
-        return response()->json(
-            $cliente->toArray()
-            );
     }
 
     /**
@@ -204,6 +113,82 @@ class ClienteController extends Controller
     public function update(Request $request, $id)
     {
         //
+        //
+        try{
+            $cliente = Cliente::find($id);
+            
+            if (!$cliente) {
+                return response("No existe el Documento", 404);
+            } 
+            $direccion = Direccion::find($cliente->direccion->id);
+            if (!$direccion) {
+                return response("No existe el Documento", 404);
+            } 
+
+            $direccion->fill($request->all());            
+            $direccion->save();  
+
+            $cliente->fill($request->all());            
+            $cliente->save();  
+
+
+            $telefono= Telefono::where('cliente_id','=',$cliente->id)->get();
+            
+
+            
+            if($telefono)
+            {   
+                $j=1;
+                for ($i=0; $i <count($telefono) ; $i++) 
+                {            
+                    $telefono[$i]['tipo']=$request->Input('telefono_tipo'.$j);
+                    $telefono[$i]['pais']=$request->Input('telefono_pais'.$j);
+                    $telefono[$i]['area']=$request->Input('telefono_area'.$j);
+                    $telefono[$i]['numero']=$request->Input('telefono_numero'.$j);
+                    
+                    $telefono[$i]->save();
+
+                    $j+=1;
+
+                                
+                }
+
+                if(count($telefono)<2)
+                {
+                    if($request->Input('telefono_tipo2')!= null)
+                    {           
+                        $telefono= new Telefono([
+                            'tipo' => $request->Input('telefono_tipo2'),      
+                            'pais' => $request->Input('telefono_pais2'),          
+                            'area' => $request->Input('telefono_area2'),          
+                            'numero' => $request->Input('telefono_numero2'),
+                            ]);
+
+                            $telefono->cliente()->associate($cliente);
+                            $telefono->save();
+                        
+                    }   
+                }
+            } 
+
+
+
+
+            //$usuario = Usuario::find(Auth::user()->id);
+            //$usuario->cliente()->attach($cliente->id);
+                      
+            
+            return response()->json(['status'=>true, 'message'=>'Muchas Gracias'], 200);
+        }
+        catch(\Exception $e)
+        {
+
+            Log::critical("No se puede actualizar el Cliente:{$e->getCode()}, {$e->getLine()}, {$e->getMessage()} ");
+            return response("Alguna cosa esta mal", 500);
+        }
+
+        
+
     }
 
     /**
@@ -214,9 +199,19 @@ class ClienteController extends Controller
      */
     public function destroy($id)
     {
-        //
-        $cliente = Cliente::find($id);
-        $cliente->delete();
-        return response()->json(['message'=>'borrado']);
+       
+        try{
+            $cliente = Cliente::find($id);
+            if (!$cliente) {
+                return response("No existe el Cliente", 404);
+            } 
+            $cliente->delete();                       
+            return response("El Cliente ha sido Eliminado", 200);
+        }
+        catch(\Exception $e)
+        {
+            Log::critical("No se puede eliminar el Cliente:{$e->getCode()}, {$e->getLine()}, {$e->getMessage()} ");
+            return response("Alguna cosa esta mal", 500);
+        }
     }
 }
